stage-controller/requirements.txt
paho-mqtt >= 1.5.1